refactor: Move TutorInfoScreen with old position to merge main
